[♻️refactor]: signIn api 함수 분리 및 배럴 export 적용
