+++ src/services/Auth/index.ts
@@ -1,0 +1,1 @@
+export { default as signIn } from './signIn';
